requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
tensorflow-cpu==2.9.1